# Benchmarks for Match; router.go cleanup