This pull request adds benchmarks for the route matching functions, and then improves on them while consolidating some of the repeated code in `router.go` into helper functions.

This is the result of `benchcmp` between the two commits in this pull request.
```
benchmark                           old ns/op     new ns/op     delta
BenchmarkRouteMatch_short           352           296           -15.91%
BenchmarkRouteMatch_long            1964          1845          -6.06%
BenchmarkRouteMatch_emptyParts      4166          3596          -13.68%
BenchmarkRouteGroupMatch_short      560           286           -48.93%
BenchmarkRouteGroupMatch_nested     55565         51172         -7.91%
```